fixed issues, refactoring
